Add license header in proper format to a Java and a Scala file for examples.
Python and other source files will be unchanged by the plugin.  This
commit changes just two files to show examples.  If this PR is
approved all the Scala and Java files in the repository (over 1,000)
must be changed to conform to the license header plugin's "preferred"
format for license headers.